--- html/movieReserveApp/src/Controller/BaseController.php
@@ -1,48 +1,0 @@
-<?php
-
-namespace App\Controller;
-
-use App\Controller\AppController;
-use Cake\Auth\DefaultPasswordHasher;
-use Cake\Event\Event;
-
-class BaseController extends AppController
-{
-
-    public function initialize()
-    {
-        parent::initialize();
-        $this->loadComponent('RequestHandler');
-        $this->loadComponent('Flash');
-        $this->loadComponent('Auth', [
-            'authorize' => ['Controller'],
-            'authenticate' => [
-                'Form' => [
-                    'fields' => ['userModel' => 'Users',
-                        'username' => 'email',
-                        'password' => 'password'
-                    ]
-                ]
-            ],
-            'loginRedirect' => [
-                'controller' => 'Mypage',
-                'action' => 'index'
-            ],
-            'logoutRedirect' => [
-                'controller' => 'Users',
-                'action' => 'login',
-            ]
-
-        ]);
-    }
-
-    public function beforeFilter(Event $event)
-    {
-        $this->Auth->allow(['add','login']);
-    }
-
-    public function isAuthorized($user)
-    {
-        return true;
-    }
-}
