+++ html/movieReserveApp/src/Controller/IndexController.php
@@ -1,19 +1,24 @@
 <?php
+
 namespace App\Controller;
 
 use App\Controller\AppController;
+use Cake\Event\Event;
 
-class IndexController extends AppController {
+class IndexController extends AppController
+{
 
-  public function initialize() {
+  public function initialize()
+  {
+    parent::initialize();
+  }
+  public function beforeFilter(Event $event)
+  {
+    $this->Auth->allow();
   }
 
-  public function index() {
-    // ビューに渡す変数
-    $title = 'QUEL CINNEMAS';
-    $login = 'ログイン';
-
-    // 変数をセットしビューに渡す
-    $this->set(compact('title', 'login'));
+  public function index()
+  {
   }
+
 }
