+++ html/movieReserveApp/src/Controller/PricesDiscountsController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
+use Cake\Event\Event;
 
 /**
  *
@@ -17,22 +18,25 @@ class PricesDiscountsController extends AppController
    */
   public function initialize()
   {
+    parent::initialize();
     $this->loadModel('RegularPrices');
     $this->loadModel('Discounts');
   }
-  public function index()
-  {
-    $title = 'QUEL CINNEMAS';
-    $login = 'ログイン';
 
-    $regular_prices = $this->RegularPrices->find('all', [
-      'conditions'=>['is_invalid'=>false],
-      'order'=>['price'=>('desc')]
-    ]);
-    $discounts = $this->Discounts->find('all', [
-      'conditions'=>['is_invalid'=>false],
-    ]);
-    $this->set(compact('title', 'login', 'regular_prices', 'discounts'));
+  public function beforeFilter(Event $event)
+  {
+    $this->Auth->allow(['index']);
   }
 
+  public function index()
+  {
+    $regular_prices = $this->RegularPrices->find('all', [
+      'conditions' => ['is_invalid' => false],
+      'order' => ['price' => ('desc')]
+    ]);
+    $discounts = $this->Discounts->find('all', [
+      'conditions' => ['is_invalid' => false],
+    ]);
+    $this->set(compact('regular_prices', 'discounts'));
+  }
 }
